New BulkDomainDescription for porous electrode.
There is an ELECTRODE_KEY_INPUT and a name for the domain
added to the constructor argument list.

This file assumes that the ELECTRODE_KEY_INPUT is fully
created in ProblemStatementCell.  It is general for anodes
cathodes, etc., as far as I can tell.

The electrolyte creation is generic with respect to liquid transport model.