Changed units on heat release to W/m3 (was J/m2)
This required adding a t_final_ and t_init_ to
m1d_porousLiIon_Separator_dom1D.  At lest that is what I added for
consistency.  Really, t_final_ and t_init_ are only used to compute
deltaTime, so we can carry around only deltaTime if we want.

Note also that there is a place where Harry set a check for rdelta_t < 1e-200
in which case those t_final_ and t_init_ will be set equal and we'll have
a divide by zero problem.

One of teh tecplot output variables, the accumulated Qsource, is still
in units of J/m3 and not Watts/m3