Improve Electrode_CSTR handling of region boundaries.
Previously if predictSoln() guessed that the solution would lie on
a region boundary but the nonlinear solve actually ended up short of
the boundary we ended up throwing in checkSubIntegrationStepAcceptable()
because of a mismatch between onRegionBoundary_final_ and RelativeExtentRxn_final_.
To prevent this we now reset onRegionBoundary_final_ during
unpackNonlinSolnVector() if the solved for extent of reaction does not
lie on the boundary.

In addition I had to update check_final_state() to update onRegionBoundary_final_
if the final solution corresponded to a region boundary so that the subsequent
subcycle step would proceed correctly.