Remade the blessed files for this test.
There is a lot more work to be done on this sample problem.
The jacobians at the end of the program don't appear to be correct.
This is associated with a problem within the global problem with
non-convergence of the global matrix.

When there is total solid state diffusion control, the jacobians
wrt voltage should go to zero. They do for most of the problem.
But when the mf in the solid get really, really tiny, this seems
to break down.

Therefore, more work must be done on this problem.