Added the stress calculation:
1) The displacement being calculated is the displacement from x0NodePos.

2) The  fluid pressure calculations have been removed are a comment substituted.
3) The stress and straing calculations have been re-activated.
	Please note that the right most residual location for stress in the Anode
	and Separator are _not_ correct residuals until the next domain completes
	the calculation. **This also means** that the residuals can _not_ be zeroed
	out for a domain on entry, as that will nuke the residual at position 0.
The calculation does run and converge for a small number of time steps. The results are
reasonable looking numbers for the noChange case, for time step one. Depending on the
values in  atolVector for displacement this failure is at step 0 to 3.

I've not tried setting the stress tollerances, that should be the first thing tried.